requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# API clients
requests>=2.31.0
semanticscholar>=0.6.0

# LLM integration
openai>=1.3.0
anthropic>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0